#3 Create new test_utilities module to hold dag_from_strings and Options class
